backend/newproject/requirements.txt
Django==5.1.6
djangorestframework==3.14.0
django-cors-headers==3.13.0
google-generativeai==0.3.1
python-dotenv==1.0.0
